requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
pillow
qrcode[pil]
channels
channels-redis
daphne
django-sslserver
reportlab
cryptography>=41.0.0
